pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docpipe"
version = "0.1.0"
description = "Unified document parsing, structured extraction, and vector ingestion pipeline"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Sunny Sinha" }]
keywords = ["document", "parsing", "extraction", "llm", "pipeline", "vector", "ingestion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "langchain-core>=0.3",
    "langchain-text-splitters>=0.3",
]

[project.optional-dependencies]
docling = ["docling>=2.0"]
langextract = ["langextract>=0.1"]
openai = ["langchain-openai>=0.3"]
google = ["langchain-google-genai>=2.0"]
ollama = ["langchain-ollama>=0.3"]
huggingface = ["langchain-huggingface>=0.1"]
pgvector = ["langchain-postgres>=0.0.12"]
server = ["fastapi>=0.100", "uvicorn[standard]>=0.20", "python-multipart>=0.0.6"]
all = ["docpipe[docling,langextract,openai,google,ollama,pgvector,server]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov",
    "ruff",
    "mypy",
    "httpx",
]

[project.scripts]
docpipe = "docpipe.cli.main:cli"

[project.entry-points."docpipe.parsers"]
docling = "docpipe.parsers.docling_parser:DoclingParser"

[project.entry-points."docpipe.extractors"]
langextract = "docpipe.extractors.langextract_extractor:LangExtractExtractor"
langchain = "docpipe.extractors.langchain_extractor:LangChainExtractor"

[project.urls]
Homepage = "https://github.com/thesunnysinha/docpipe"
Repository = "https://github.com/thesunnysinha/docpipe"
Issues = "https://github.com/thesunnysinha/docpipe/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/docpipe"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "requires_docling: needs docling installed",
    "requires_langextract: needs langextract installed",
    "requires_langchain: needs langchain provider installed",
    "requires_pgvector: needs pgvector DB available",
    "requires_api_key: needs LLM API key configured",
]
